backend/requirements.txt
flask
ultralytics      # gives you YOLO loader + inference
Pillow           # for image decoding
